functions logicas front detail
