+++ client/src/Components/Details/LogicFunctions/ContratosVisibles.js
